move shared fixtures
