+++ cumulusci/core/keychain/tests/conftest.py
@@ -1,0 +1,39 @@
+import pytest
+
+from cumulusci.core.config import (
+    UniversalConfig,
+    BaseProjectConfig,
+    OrgConfig,
+    ServiceConfig,
+)
+
+
+@pytest.fixture()
+def project_config():
+    universal_config = UniversalConfig()
+    project_config = BaseProjectConfig(universal_config, config={"no_yaml": True})
+    project_config.config["services"] = {
+        "connected_app": {"attributes": {"test": {"required": True}}},
+        "github": {"attributes": {"name": {"required": True}, "password": {}}},
+        "not_configured": {"attributes": {"foo": {"required": True}}},
+        "devhub": {"attributes": {"foo": {"required": True}}},
+    }
+    project_config.project__name = "TestProject"
+    return project_config
+
+
+@pytest.fixture
+def org_config():
+    return OrgConfig({"foo": "bar"}, "test")
+
+
+@pytest.fixture
+def key():
+    return "0123456789123456"
+
+
+@pytest.fixture
+def service_config():
+    return ServiceConfig(
+        {"name": "bar@baz.biz", "password": "test123", "token": "test123"}
+    )
